+++ lib/src/tangent_sdk.dart
@@ -285,7 +285,7 @@ class TangentSDK {
 
   Future<Result<CustomerPurchasesInfo>> restorePurchases() async {
     return await _revenueService?.restorePurchases() ??
-        Success(CustomerPurchasesInfo(hasActiveSubscription: false, originalAppUserId: '', purchases: []));
+        Success(CustomerPurchasesInfo(hasActiveSubscription: false, originalAppUserId: '', purchases: [], managementURL: null));
   }
 
   Future<Result<void>> logIn(String appUserId) async {
@@ -302,6 +302,10 @@ class TangentSDK {
 
   Stream<CustomerPurchasesInfo> get customerPurchasesInfoStream =>
       _revenueService?.customerPurchasesInfoStream ?? const Stream.empty();
+
+  Future<Result<String?>> getManagementUrl() async {
+    return await _revenueService?.getManagementUrl() ?? const Success(null);
+  }
 
   /// Determines if the given `productId` is a renewal purchase based on the
   /// existing purchase history.